refactor: extract hardcoded content arrays into src/content/
Move bot status pools, amikool reply pools, friendship-listener phrase
triggers, notice category styling, and accolade/achievement display
metadata out of the services that consume them and into a new
`src/content/` folder so contributors can edit copy without touching
service logic.

Accolade/achievement awarding logic stays in `achievements-service.ts`,
keyed by the same id as the metadata. The `Record<AccoladeType, ...>`
typing makes adding metadata-without-logic (or vice versa) a compile
error.

No behavior change. Includes a README describing what belongs in
`src/content/` and what doesn't (e.g. runtime-configurable values still
belong in ConfigService).

Refs https://github.com/lonix/koolbot/issues/365
